Upgrade from Chromium 89 to Chromium 90 (1.23.x).

+++ browser/component_updater/brave_component_updater_configurator.cc
@@ -57,7 +57,7 @@ BraveConfigurator::BraveConfigurator(
 
 BraveConfigurator::~BraveConfigurator() {}
 
-int BraveConfigurator::InitialDelay() const {
+double BraveConfigurator::InitialDelay() const {
   return configurator_impl_.InitialDelay();
 }
